fixed a bug in `getDateRelative()`
I have modified `getDateRelative()` so that the basis for calculating the time difference is the date and not the time (86400s (24h)). This had the consequence that `getDateRelative()` a page that was created yesterday at 18:00 o'clock, today at 13:00 o'clock `today` output, which is definitely wrong. Now the date of the publication is checked and compared with today's date.